Rename shared folder, add to components
