+++ src/Patient/ForcePasswordChange.js
@@ -3,7 +3,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import useStores from '../Basics/UseStores';
 import { observer } from 'mobx-react'
 import { useTranslation } from 'react-i18next';
-import PasswordUpdate from '../Shared/PasswordUpdate'
+import PasswordUpdate from '../Components/PasswordUpdate'
 import ClickableText from '../Basics/ClickableText';
 import Typography from '@material-ui/core/Typography';
 import Exit from '@material-ui/icons/ExitToApp';
